fix docstring

+++ specialk/models/mt_model.py
@@ -149,7 +149,7 @@ class NMTModule(pl.LightningModule):
         """Calculates BLEU score @ validation phase.
 
         Args:
-            y_hat (Int[Tensor, "batch seq_len vocab"]): Prediction tensor.
+            y_hat (Float[Tensor, "batch seq_len vocab"]): Prediction tensor.
             y (Int[Tensor, "batch seq_len"]): Reference tensor.
 
         Returns:
